# parser/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dmarc-parser API",
    "version": "2.0.0",
    "description": "REST API of the DMARC aggregate report parser & ingestion service.\n\n### Authentication\nEvery `/api/v1/*` route (except `/api/v1/openapi.json`) requires an API\nkey from `PARSER_API_KEYS`, sent either as `Authorization: Bearer <key>`\nor `X-API-Key: <key>`. Keys carry scopes from `{read, ingest, admin}`;\neach operation documents its required scope via `x-required-scope`.\nMissing/invalid key \u2192 401; key without the required scope \u2192 403.\n\n### Rate limiting\nPer-key token bucket (`PARSER_RATE_LIMIT_RPS`, default 10 req/s,\nburst 3x). Exceeding it returns 429 with a `Retry-After` header.\n\n### Conventions\nAll timestamps are UTC, RFC 3339. Date-only values accepted by `since` /\n`until` parameters use `YYYY-MM-DD`. Errors are JSON\n`{\"error\": \"message\"}`.\n"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Host loopback"
    },
    {
      "url": "http://dmarcparser:8080",
      "description": "From containers on the mxsentinel_default / dmarcparser_default networks"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/reports/{serial}/raw": {
      "get": {
        "tags": [
          "reports",
          "export"
        ],
        "summary": "Download the raw report XML",
        "description": "404 when the serial is unknown or raw_xml was purged by the retention job (PARSER_RAW_XML_RETENTION).\n",
        "operationId": "getReportRaw",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/serial"
          }
        ],
        "responses": {
          "200": {
            "description": "Raw XML attachment",
            "content": {
              "text/xml": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/export": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Stream flattened report\u00d7record rows (CSV or JSONL)",
        "description": "Chunked streaming export of every record row matching the filters, joined with its report metadata. CSV has a header row; JSONL emits one ExportRow object per line.\n",
        "operationId": "exportRecords",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "jsonl"
              ],
              "default": "csv"
            }
          },
          {
            "$ref": "#/components/parameters/domain"
          },
          {
            "$ref": "#/components/parameters/org"
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/until"
          }
        ],
        "responses": {
          "200": {
            "description": "Streamed export",
            "content": {
              "text/csv": {},
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRow"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Query the API audit trail",
        "operationId": "listAudit",
        "x-required-scope": "admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "name": "actor",
            "in": "query",
            "description": "Filter by API key name",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/limit1000"
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entries, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditEntry"
                      }
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/requeue-failed": {
      "post": {
        "tags": [
          "poller"
        ],
        "summary": "Requeue all mail from the Failed folder",
        "description": "Moves every message in the Failed IMAP folder back to INBOX (cleared of \\Seen) and triggers an immediate poll cycle.\n",
        "operationId": "requeueFailed",
        "x-required-scope": "admin",
        "responses": {
          "200": {
            "description": "Requeue result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "requeued": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Poller disabled (PARSER_IMAP_ADDR empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "502": {
            "description": "IMAP operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/replay": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Re-deliver dead-lettered webhooks",
        "description": "Attempts one synchronous delivery for every un-replayed webhook_deadletter row (oldest first, max 500 per call); successful deliveries are marked replayed.\n",
        "operationId": "replayWebhooks",
        "x-required-scope": "admin",
        "responses": {
          "200": {
            "description": "Replay result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pending": {
                      "type": "integer",
                      "description": "Rows attempted in this call"
                    },
                    "replayed": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/digest/subscriptions": {
      "get": {
        "tags": [
          "digest"
        ],
        "summary": "List digest subscriptions",
        "operationId": "listDigestSubscriptions",
        "x-required-scope": "admin",
        "responses": {
          "200": {
            "description": "Subscriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscriptions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DigestSubscription"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "post": {
        "tags": [
          "digest"
        ],
        "summary": "Subscribe an email address to a domain's weekly digest",
        "operationId": "addDigestSubscription",
        "x-required-scope": "admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "domain",
                  "email"
                ],
                "properties": {
                  "domain": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestSubscription"
                }
              }
            }
          },
          "200": {
            "description": "Already existed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestSubscription"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/digest/subscriptions/{id}": {
      "delete": {
        "tags": [
          "digest"
        ],
        "summary": "Delete a digest subscription",
        "operationId": "deleteDigestSubscription",
        "x-required-scope": "admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "deleted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/digest/run": {
      "post": {
        "tags": [
          "digest"
        ],
        "summary": "Force-send all digests now",
        "description": "Sends a digest for every subscription immediately, bypassing the weekly schedule guard. Intended for testing.",
        "operationId": "runDigest",
        "x-required-scope": "admin",
        "responses": {
          "200": {
            "description": "Run summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestRunSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Run failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from PARSER_API_KEYS as a bearer token. Key scopes: read, ingest, admin (see each operation's x-required-scope)."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same API key via the X-API-Key header."
      }
    },
    "parameters": {
      "domain": {
        "name": "domain",
        "in": "query",
        "description": "Filter by policy-published domain",
        "schema": {
          "type": "string"
        }
      },
      "org": {
        "name": "org",
        "in": "query",
        "description": "Filter by reporting organisation",
        "schema": {
          "type": "string"
        }
      },
      "since": {
        "name": "since",
        "in": "query",
        "description": "Lower time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
      "until": {
        "name": "until",
        "in": "query",
        "description": "Upper time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
...
      "limit1000": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 100
        }
      },
...
      "serial": {
        "name": "serial",
        "in": "path",
        "required": true,
        "description": "Report serial",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Per-key rate limit exceeded (see Retry-After)",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ExportRow": {
        "type": "object",
        "description": "One flattened report\u00d7record row (JSONL export; CSV columns match in order)",
        "properties": {
          "serial": {
            "type": "integer",
            "format": "int64"
          },
          "domain": {
            "type": "string"
          },
          "org": {
            "type": "string"
          },
          "report_id": {
            "type": "string"
          },
          "date_begin": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "date_end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "source_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "disposition": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_align": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_align": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_result": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "spf_result": {
            "type": [
              "string",
              "null"
            ]
          },
          "header_from": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AuditEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ts": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "type": [
              "string",
              "null"
            ],
            "description": "API key name"
          },
          "action": {
            "type": [
              "string",
              "null"
            ],
            "description": "HTTP method"
          },
          "route": {
            "type": [
              "string",
              "null"
            ],
            "description": "Route pattern"
          },
          "status": {
            "type": [
              "integer",
              "null"
            ]
          },
          "client_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "description": "Extra context, e.g. {\"serials\":[...]} for ingests",
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "DigestSubscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "domain": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "DigestRunSummary": {
        "type": "object",
        "properties": {
          "domains_sent": {
            "type": "integer"
          },
          "domains_skipped": {
            "type": "integer"
          },
          "emails_sent": {
            "type": "integer"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}